mhealth-voice-job/requirements.txt
alembic==1.9.3
amqp==5.1.1
anyio==3.6.2
appdirs==1.4.4
async-timeout==4.0.2
bcrypt==4.0.1
billiard==3.6.4.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==41.0.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.92.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
humanize==4.6.0
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.2.4
loguru==0.6.0
Mako==1.2.4
MarkupSafe==2.1.2
orjson==3.8.6
packaging==23.1
passlib==1.7.4
prometheus-client==0.16.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.4
PyMySQL==1.0.3
python-dateutil==2.8.2
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-toolbelt==1.0.0
rfc3986==1.5.0
rsa==4.9
s3transfer==0.6.1
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.3
starlette==0.25.0
tornado==6.2
tqdm==4.65.0
typing_extensions==4.4.0
ujson==5.7.0
urllib3==1.26.14
uvicorn==0.20.0
vine==5.0.0
watchfiles==0.18.1
wcwidth==0.2.6
win32-setctime==1.1.0
